Re-factoring models to closer match current layout
