+++ app/gwells/forms/ActivitySubmissionGpsForm.py
@@ -22,8 +22,7 @@ from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Layout, Fieldset, Div, Submit, Hidden, HTML, Field
 from crispy_forms.bootstrap import FormActions, AppendedText, InlineRadios
 
-from gwells.models import ActivitySubmission
-from wells.models import DrillingMethodCode
+from wells.models import DrillingMethodCode, ActivitySubmission
 
 
 logger = logging.getLogger(__name__)
