+++ app/gwells/forms/ActivitySubmissionWaterQualityForm.py
@@ -19,7 +19,7 @@ from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Layout, Fieldset, Div, Submit, Hidden, HTML, Field
 from crispy_forms.bootstrap import FormActions, AppendedText, InlineRadios
 
-from gwells.models import ActivitySubmission
+from wells.models import ActivitySubmission
 
 
 class ActivitySubmissionWaterQualityForm(forms.ModelForm):
