+++ app/gwells/forms/LithologyForm.py
@@ -22,7 +22,7 @@ from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Layout, Fieldset, Div, Submit, Hidden, HTML, Field
 from crispy_forms.bootstrap import FormActions, AppendedText, InlineRadios
 
-from gwells.models import LithologyDescription
+from wells.models import LithologyDescription
 
 logger = logging.getLogger(__name__)
