+++ app/gwells/forms/ScreenForm.py
@@ -19,7 +19,7 @@ from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Layout, Fieldset, Div, Submit, Hidden, HTML, Field
 from crispy_forms.bootstrap import FormActions, AppendedText, InlineRadios
 
-from gwells.models import Screen
+from wells.models import Screen
 
 
 class ScreenForm(forms.ModelForm):
