+++ app/gwells/forms/WellOwnerForm.py
@@ -21,7 +21,7 @@ from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Layout, Fieldset, Div, Submit, Hidden, HTML, Field
 from crispy_forms.bootstrap import FormActions, AppendedText, InlineRadios
 
-from gwells.models import ActivitySubmission
+from wells.models import ActivitySubmission
 
 
 class WellOwnerForm(forms.ModelForm):
