+++ app/gwells/views/health_view.py
@@ -1,7 +1,7 @@
 from django.views import generic
 from django.http import HttpResponse
 
-from gwells.models import WellActivityCode
+from submissions.models import WellActivityCode
 
 
 class HealthView(generic.TemplateView):
