+++ app/submissions/serializers.py
@@ -13,11 +13,9 @@
 """
 
 from rest_framework import serializers
+
 from gwells.models import ProvinceStateCode
-from wells.models import (
-    Well,
-    ActivitySubmission
-)
+from wells.models import Well, ActivitySubmission
 from wells.stack import StackWells
 from gwells.serializers import AuditModelSerializer
